+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,7 +1,7 @@
 use crate::jobs::import_groups::finalize_group_import;
 use crate::lifecycle::{init_env, init_state};
-use crate::memory::{get_chat_events_memory, get_upgrades_memory};
-use crate::{read_state, Data};
+use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
+use crate::{mutate_state, read_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use community_canister::post_upgrade::Args;
@@ -14,7 +14,7 @@ use types::CanisterId;
 #[post_upgrade]
 #[trace]
 fn post_upgrade(args: Args) {
-    chat_events::ChatEvents::init_stable_storage(get_chat_events_memory());
+    stable_memory_map::init(get_stable_memory_map_memory());
 
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
@@ -53,5 +53,12 @@ fn post_upgrade(args: Args) {
         state
             .data
             .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now)
+    });
+
+    mutate_state(|state| {
+        let now = state.env.now();
+        for channel in state.data.channels.iter_mut() {
+            channel.chat.events.remove_spurious_video_call_in_progress(now);
+        }
     });
 }