+++ backend/canisters/group/impl/src/updates/toggle_mute_notifications.rs
@@ -2,7 +2,6 @@ use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use group_canister::toggle_mute_notifications::{Response::*, *};
-use types::Timestamped;
 
 #[update(candid = true, msgpack = true)]
 #[trace]
@@ -15,13 +14,15 @@ fn toggle_mute_notifications(args: Args) -> Response {
 fn toggle_mute_notifications_impl(args: Args, state: &mut RuntimeState) -> Response {
     let caller = state.env.caller();
     let now = state.env.now();
-    match state.data.get_member_mut(caller) {
-        Some(member) => {
-            member.notifications_muted = Timestamped::new(args.mute, now);
-            let user_id = member.user_id();
+    if let Some(user_id) = state.data.lookup_user_id(caller) {
+        if matches!(
+            state.data.chat.members.toggle_notifications_muted(user_id, args.mute, now),
+            Some(true)
+        ) {
             state.data.mark_group_updated_in_user_canister(user_id);
-            Success
         }
-        None => CallerNotInGroup,
+        Success
+    } else {
+        CallerNotInGroup
     }
 }