+++ backend/canisters/user/impl/src/lifecycle/post_upgrade.rs
@@ -1,6 +1,6 @@
 use crate::lifecycle::{init_env, init_state};
-use crate::memory::{get_chat_events_memory, get_upgrades_memory};
-use crate::Data;
+use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
+use crate::{mutate_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
@@ -12,7 +12,7 @@ use user_canister::post_upgrade::Args;
 #[post_upgrade]
 #[trace]
 fn post_upgrade(args: Args) {
-    chat_events::ChatEvents::init_stable_storage(get_chat_events_memory());
+    stable_memory_map::init(get_stable_memory_map_memory());
 
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
@@ -34,4 +34,11 @@ fn post_upgrade(args: Args) {
     init_state(env, data, args.wasm_version);
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
+
+    mutate_state(|state| {
+        let now = state.env.now();
+        for chat in state.data.direct_chats.iter_mut() {
+            chat.events.remove_spurious_video_call_in_progress(now);
+        }
+    });
 }