+++ backend/canisters/user_index/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Pass in `BotApiCanister` when installing a new LocalUserIndex ([#6828](https://github.com/open-chat-labs/open-chat/pull/6828))
 - Simplify `inspect_message` ([#6847](https://github.com/open-chat-labs/open-chat/pull/6847))
 - Allow bots to set a display name when registering ([#6850](https://github.com/open-chat-labs/open-chat/pull/6850))
+- Remove diamond only access to set display name ([#6879](https://github.com/open-chat-labs/open-chat/pull/6879))
 
 ## [[2.0.1450](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1450-user_index)] - 2024-11-14